Move programd to core (and rename)
