+++ jetson/programd.sh
@@ -1,3 +1,3 @@
 #!/bin/bash -ex
 . setup.bash
-python -m farm_ng.program_supervisor
+python -m farm_ng.core.programd
